refactor: refactor responders
