+++ file-manager/modules/cli/index.js
@@ -1,5 +1,5 @@
 export { args } from "./args/args.js";
 export { colorizer } from "./utils/colorizer.js";
 export { cliDrawer } from "./utils/cliDrawer.js";
-export { responders } from "./responders/responders.js";
+export * as responders from "./responders/index.js";
 export { workingDir } from "./utils/workingDir.js";
