+++ file-manager/modules/cli/responders/exit/exit.js
@@ -1,0 +1,9 @@
+/*
+ * CLI responder exit
+ *
+ */
+
+// Exiting CLI application
+export const exitResponder = () => {
+  process.exit(0);
+};
