+++ file-manager/modules/cli/responders/index.js
@@ -1,0 +1,5 @@
+export { exitResponder as exit } from "./exit/exit.js";
+export { cdResponder as cd } from "./cd/cd.js";
+export { hashResponder as hash } from "./hash/hash.js";
+export { osResponder as os } from "./os/os.js";
+export { upResponder as up } from "./up/up.js";
